Split out tests into separate file
